AIVENTS-Chatbot-ready-for-tests/requirements.txt
langchain<=0.2
langchain-community
langchain-openai
openai
pyodbc
azure-identity
langgraph
networkx
langchainhub
matplotlib
pygraphviz
playwright
openpyxl
pandas
pyyaml
langid
rapidfuzz
unidecode
nltk
stanza
spacy
unidecode
nbformat
ragas
datacompy
python-dotenv
bert_score
XlsxWriter
nest_asyncio
ragas
sqlalchemy
sqlalchemy-pytds
python-tds
streamlit